+ Added basic shell

+++ src/libc/syscall.h
@@ -5,11 +5,11 @@
 #include <stdint.h>
 
 // To call a syscall without setting parameters to 0
-#define SYSC0(ID)                           syscall(ID, 0, 0, 0, 0)
-#define SYSC1(ID, arg1)                     syscall(ID, arg1, 0, 0, 0)
-#define SYSC2(ID, arg1, arg2)               syscall(ID, arg1, arg2, 0, 0)
-#define SYSC3(ID, arg1, arg2, arg3)         syscall(ID, arg1, arg2, arg3, 0)
-#define SYSC4(ID, arg1, arg2, arg3, arg4)   syscall(ID, arg1, arg2, arg3, arg4)
+#define SYSC0(ID)                           syscall((uint32_t)ID, 0, 0, 0, 0)
+#define SYSC1(ID, arg1)                     syscall((uint32_t)ID, (uint32_t)arg1, 0, 0, 0)
+#define SYSC2(ID, arg1, arg2)               syscall((uint32_t)ID, (uint32_t)arg1, (uint32_t)arg2, 0, 0)
+#define SYSC3(ID, arg1, arg2, arg3)         syscall((uint32_t)ID, (uint32_t)arg1, (uint32_t)arg2, (uint32_t)arg3, 0)
+#define SYSC4(ID, arg1, arg2, arg3, arg4)   syscall((uint32_t)ID, (uint32_t)arg1, (uint32_t)arg2, (uint32_t)arg3, (uint32_t)arg4)
 
 // Syscalls ids
 // See doc/syscalls.md for specs
